requirements.txt
python-pptx==1.0.2
loguru==0.7.2
gradio==5.1.0
langchain==0.2.16
langchain_core==0.2.41
langchain_community==0.2.17
langchain_ollama==0.1.3
langchain_openai==0.1.25
python-docx==1.1.2
Pillow==10.3.0
torch==2.5.0
transformers==4.46.0
datasets==3.0.2
accelerate==1.0.1
librosa==0.10.2.post1
soundfile==0.12.1
ffmpeg==1.4
torchvision==0.20.0
sentencepiece==0.2.0
bitsandbytes==0.44.1